pyproject.toml
[project]
name = "edinetanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "edinet-xbrl>=0.2.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "pandas>=2.3.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "psutil>=7.0.0",
    "pytest-mock>=3.14.1",
    "requests-mock>=1.12.1",
]
